@@ -1,12 +1,11 @@
 import { Component } from '@angular/core';
+import { RouterModule } from '@angular/router';
 
 @Component({
   selector: 'app-cookies',
   standalone: true,
-  imports: [],
+  imports: [RouterModule],
   templateUrl: './cookies.component.html',
-  styleUrl: './cookies.component.scss'
+  styleUrl: './cookies.component.scss',
 })
-export class CookiesComponent {
-
-}
+export class CookiesComponent {}
